# swagger.yaml
openapi: 3.0.3
info:
  title: Meraki API
  description: Meraki Backend API
  version: 1.0.0
servers:
  - url: http://localhost:4000
...
paths:
...
  /favorites/{userId}:
    get:
      tags:
        - favorites
      summary: Get Favorites
      description: Get all favorite pokemons from a user
      operationId: getFavorites
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PokemonResponseDTO'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
...
components:
  schemas:
...
    PokemonResponseDTO:
      type: object
      properties:
        data:
          type: object
          properties:
            uuid:
              type: string
            name:
              type: string
            type:
              type: string
            nickname:
              type: string
            level:
              type: integer
            userId:
              type: string
        messages:
          type: array
          items:
            type: string
        success:
          type: boolean
    #ERROR SCHEMA
    ErrorDTO:
      type: object
      properties:
        data:
          type: object
        messages:
          type: array
          items:
            type: string
        success:
          type: boolean